delete now returns the deleted object

+++ routes/posts-router.js
@@ -49,18 +49,33 @@ router.post('/', (req , res) => {
 
 router.delete('/:id', (req, res) => {
     const { id } = req.params
-    db.remove(id)
-        .then( del => {
-            if (!del) {
-                res.status(404).json({ message: "The post with the specified ID does not exist." })
+    db.findById(id)
+        .then(post => {
+            if (post[0]) {
+                db.remove(id)
+                    .then( ()  => {
+                        res.status(200).json(post[0])
+                    })
             }
             else {
-                res.status(200).json(del)
+                res.status(404).json({ message: "The post with the specified ID does not exist."})
             }
         })
-        .catch( err => {
+        .catch(err => {
             res.status(500).json({ error: "The post could not be removed" })
-        })
+        })  
+
+        // .then( del => {
+        //     if (!del) {
+        //         res.status(404).json({ message: "The post with the specified ID does not exist." })
+        //     }
+        //     else {
+        //         res.status(200).json(del)
+        //     }
+        // })
+        // .catch( err => {
+        //     res.status(500).json({ error: "The post could not be removed" })
+        // })
     
 })
